Integration testing (#7)
* Tested AuthController

* Tested UserController

* Tully tested CompanyController

* Tested ReportController

+++ src/main/java/com/bond/service/CompanyServiceImpl.java
@@ -56,7 +56,9 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public void deleteById(UUID id) {
-        companyRepository.deleteById(id);
+        if (companyRepository.findById(id).isPresent()) {
+            companyRepository.deleteById(id);
+        }
     }
 
     @Override
