+++ src/main/java/com/bond/service/UserServiceImpl.java
@@ -83,7 +83,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        if (userRepository.findById(id).isPresent()) {
+            userRepository.deleteById(id);
+        }
     }
 
     private boolean userIs(User user, Role.RoleName roleName) {
